Revised signal fastpath
Brought in more scheduler logic to the
signal fastpath so that signaling to a
higher prio thread more closely follows
the slowpath. Cleaned up the code to
avoid repetition and unnecessary code.
Modified to be safe i.e. no state changes
until after the point of no return.
Removed support for non-AARCH64 platforms.

Signed-off-by: Alwin Joshy <a.joshy@student.unsw.edu.au>