Removed signal to higher prio path
After the discussion on the PR, it was decided that the best course
of action was to only keep the signal to low prio threads fastpath.
Removed the signal to higher prio fastpath and all it used
that is no longer necessary. Also fixed a cmake misconfiguration
that previously made it annoying to switch the signal fastpath on
and off. Fixed guards around signal fastpath so it wouldn't
stop other configurations from compiling and did some minor style
fixes.

Signed-off-by: Alwin Joshy <joshyalwin@gmail.com>