wip: AARCH64 hardware debug API
I am beginning to add support for AARCH64 in the hardware debug API. This is quite
painful because the existing codebase was designed many years ago and seems to have
been written in a way that is specific to ARMv7. Most of what I've done so far
is roughly shuffling things around to get them into a more sane configuration before
I implement the API and properly clean things up.

Signed-off-by: Alwin Joshy <joshyalwin@gmail.com>