HW Debug API: Better handling of BAS on ARM
The watchpoint value register contains a 4-byte aligned virtual address,
but the way we set the BAS, which determines the bytes of the word
pointed to by the address that trigger an exception, assumes the
address can be anything. This means that watchpoints did not reliably
work for small types such as characters and shorts, which are not
always 4-byte aligned.  In this patch, if a watchpoint with size < 4
is set, we align the address used to the previous 4-byte boundary and
set the appropriate BAS to trigger an exception for the sub-range
intended by the user.

Signed-off-by: Alwin Joshy <joshyalwin@gmail.com>